# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Empty Spain API",
        "contact": {
            "name": "Grupo 5 Support Team",
            "email": "grupo5@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Public access"
                ],
                "summary": "Checks if the user is in the database",
                "operationId": "App\\Http\\Controllers\\AuthController::authenticate",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Ronaldo@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Email or password invalid"
                    },
                    "200": {
                        "description": "Success since the user credentials are correct."
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Public access"
                ],
                "summary": "Register a user",
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surnames",
                                    "password",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "surnames": {
                                        "$ref": "#/components/schemas/User/properties/surnames"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    },
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validator fail such as email is not unique"
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/addSearch": {
            "post": {
                "tags": [
                    "Public access"
                ],
                "summary": "Add search info after a user searches a municipio",
                "operationId": "App\\Http\\Controllers\\BusquedaController::addSearch",
                "requestBody": {
                    "description": "JSON containing scraper info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tripadvisor_info",
                                    "twitter_info",
                                    "tiempo_info",
                                    "wiki_info",
                                    "municipio_id",
                                    "usuario_id",
                                    "municipio_state"
                                ],
                                "properties": {
                                    "tripadvisor_info": {
                                        "$ref": "#/components/schemas/Busqueda/properties/tripadvisor_info"
                                    },
                                    "twitter_info": {
                                        "$ref": "#/components/schemas/Busqueda/properties/twitter_info"
                                    },
                                    "tiempo_info": {
                                        "$ref": "#/components/schemas/Busqueda/properties/tiempo_info"
                                    },
                                    "wiki_info": {
                                        "$ref": "#/components/schemas/Busqueda/properties/wiki_info"
                                    },
                                    "municipio_id": {
                                        "$ref": "#/components/schemas/Busqueda/properties/municipio_id"
                                    },
                                    "usuario_id": {
                                        "$ref": "#/components/schemas/Busqueda/properties/usuario_id"
                                    },
                                    "municipio_state": {
                                        "$ref": "#/components/schemas/Busqueda/properties/municipio_state"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/municipios/{id}": {
            "get": {
                "tags": [
                    "Public access"
                ],
                "summary": " Gets information belonging to a municipio",
                "operationId": "App\\Http\\Controllers\\BusquedaController::municipioInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the municipio",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "2045"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}